LayoutTests/platform/chromium-win/http/tests/eventsource/existent-eventsource-status-error-iframe-crash-expected.txt

CONSOLE MESSAGE: line 1: EventSource's response has a MIME type ("text/html") that is not "text/event-stream". Aborting the connection.
PASS
